update evaluation endpoint

+++ src/main/java/fr/ubo/dosi/CSCIEVAE/services/EvaluationServiceImpl.java
@@ -187,12 +187,17 @@ public class EvaluationServiceImpl implements EvaluationService{
     @Override
     public EvaluationDTO updateRubriquesEvaluationOrder(EvaluationDTO evaluationDTO) {
         log.info(" __ Updates sur les rubriques associées aux évaluations __ ");
-           rubriqueEvalutionRepository.deleteAllByIdEvaluation(evaluationDTO.getIdEvaluation());
-           Evaluation evaluation = dataMapper.evaluationDtoToEvaluation(evaluationDTO);
+        if (!rubriqueEvalutionRepository.findAllByIdEvaluationOrderByOrdreAsc(evaluationDTO.getIdEvaluation())
+                .isEmpty()){
+            log.info("  ___ Suppression des Rub Eva associer à l'évaluation ___");
+            rubriqueEvalutionRepository.deleteAllByIdEvaluation(evaluationDTO.getIdEvaluation());
+        }else {
+            Evaluation evaluation = dataMapper.evaluationDtoToEvaluation(evaluationDTO);
            log.info(" -- Associer l'evaluation au nouveaux rubriques -- ");
            associetRubriquesToEvaluation(evaluation, evaluationDTO.getRubriques());
            evaluationDTO.setRubriques(getRubriqueEvaluation(evaluation.getIdEvaluation()));
            log.info("-- Evaluation updated successfully --");
+        }
         return evaluationDTO;
     }
